+++ lotus_backend/src/main/java/isamrs/tim17/lotus/model/User.java
@@ -20,7 +20,7 @@ import javax.persistence.InheritanceType;
 		)
 public abstract class User {
 	@Id
-	private int id;
+	private long id;
 	@Column(name = "email", unique = true, nullable = false)
 	private String email;
 	@Column(name = "password", unique = false, nullable = false)
@@ -45,7 +45,7 @@ public abstract class User {
 	public User() {}
 	
 	public User(String email, String password, String name, String surname, String address, String city, String country,
-			String phoneNumber, int id, UserType type, Date birthDate, Gender gender) {
+			String phoneNumber, long id, UserType type, Date birthDate, Gender gender) {
 		super();
 		this.email = email;
 		this.password = password;
@@ -124,11 +124,11 @@ public abstract class User {
 		this.phoneNumber = phoneNumber;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
